requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.21.0
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
evaluate==0.4.3
expecttest==0.2.1
faiss-cpu==1.8.0.post1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.65.0
grad-cam==1.5.3
greenlet==3.0.3
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hypothesis==6.98.10
jiter==0.5.0
joblib==1.4.2
JPype1==1.5.0
jq==1.8.0
jsonpatch==1.33
jsonpointer==2.1
kiwipiepy==0.18.0
kiwipiepy-model==0.18.0
kiwisolver==1.4.7
konlpy==0.6.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-experimental==0.0.65
langchain-huggingface==0.0.3
langchain-openai==0.1.23
langchain-teddynote==0.0.30
langchain-text-splitters==0.2.4
langchain-upstage==0.2.0
langsmith==0.1.116
lightning==2.4.0
lightning-utilities==0.11.7
lxml==5.3.0
lz4==4.3.3
marshmallow==3.22.0
matplotlib==3.9.2
mkl-service==2.4.0
mmh3==4.1.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nltk==3.9.1
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
olefile==0.47
openai==1.44.0
opencv-python==4.10.0.84
optree==0.10.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow 
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
pyarrow==17.0.0
pydantic==2.9.0
pydantic_core==2.23.2
pyparsing==3.1.4
pypdf==4.3.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python-etcd==0.4.5
pytorch-lightning==2.4.0
rank-bm25==0.2.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.0.1
sentry-sdk==2.14.0
setproctitle==1.3.3
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.19.1
torch==2.2.1
torchaudio==2.2.1
torchelastic==0.2.2
torchmetrics==1.4.2
torchvision==0.17.1
tqdm==4.66.5
transformers==4.44.2
triton==2.2.0
ttach==0.0.3
types-dataclasses==0.6.6
types-requests==2.32.0.20240907
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
umap==0.1.1
wandb==0.18.1
wget==3.2
xxhash==3.5.0
yarl==1.10.0